Feat: SCREEN_ORIENTATION_PORTRAIT
