--- app/src/main/java/com/sesac/lactosefree/common/DefaultApplication.kt
@@ -1,15 +1,0 @@
-package com.sesac.lactosefree.common
-
-import android.app.Application
-
-class DefaultApplication: Application() {
-    init {
-        instance = this
-    }
-    companion object {
-        private var instance: DefaultApplication? = null
-        fun applicationContext() : DefaultApplication {
-            return instance as DefaultApplication
-        }
-    }
-}
