+++ app/src/main/java/com/yoon/lactosefree/common/DefaultApplication.kt
@@ -1,0 +1,33 @@
+package com.yoon.lactosefree.common
+
+import android.annotation.SuppressLint
+import android.app.Activity
+import android.app.Application
+import android.content.pm.ActivityInfo
+import android.os.Bundle
+
+class DefaultApplication: Application() {
+    companion object {
+        private var instance: DefaultApplication? = null
+        fun applicationContext() : DefaultApplication {
+            return instance as DefaultApplication
+        }
+    }
+    override fun onCreate() {
+        super.onCreate()
+        instance = this
+        registerActivityLifecycleCallbacks(object : ActivityLifecycleCallbacks {
+            @SuppressLint("SourceLockedOrientationActivity")
+            override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
+                activity?.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT;
+            }
+            override fun onActivityStarted(activity: Activity) {}
+            override fun onActivityResumed(activity: Activity) {}
+            override fun onActivityPaused(activity: Activity) {}
+            override fun onActivityStopped(activity: Activity) {}
+            override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {}
+            override fun onActivityDestroyed(activity: Activity) {}
+
+        })
+    }
+}
